add loading image animation and LQIP feature
